requirements.txt
fastapi==0.110.1
uvicorn==0.17.6
pydantic==2.7.0
PyYAML==6.0
natasha==1.4.0
# pullenti==3.2
# pymorphy2